requirements.txt
streamlit>=1.46.0
pandas>=2.3.0
numpy>=2.3.0
requests>=2.32.0
python-dotenv>=1.0.0
openai>=1.0.0
plotly>=5.17.0
altair>=5.5.0
Pillow>=11.2.0
python-dateutil>=2.9.0 